+++ driver-core/CHANGELOG.rst
@@ -7,6 +7,7 @@ CHANGELOG
 - [improvement] Shade Netty dependency (JAVA-538)
 - [improvement] Target schema refreshes more precisely (JAVA-543)
 - [new feature] Add AddressTranslater for EC2 multi-region deployment (JAVA-518)
+- [bug] Don't check rpc_address for control host (JAVA-546)
 
 
 2.0.8: